datagenerator/arm_robot_machine_sensor/requirements.txt
kafka-python==2.0.5